+++ source/B2B.Transactions.CimMessageAdapter/MessageReceiver.cs
@@ -23,7 +23,6 @@ using B2B.CimMessageAdapter.Transactions;
 using B2B.Transactions;
 using B2B.Transactions.Authentication;
 using B2B.Transactions.IncomingMessages;
-using B2B.Transactions.Messages;
 using B2B.Transactions.Transactions;
 using B2B.Transactions.Xml.Incoming;
