+++ source/B2B.Transactions.CimMessageAdapter/Messages/MessageParser.cs
@@ -19,7 +19,7 @@ using System.Threading.Tasks;
 using System.Xml;
 using System.Xml.Schema;
 using B2B.CimMessageAdapter.Errors;
-using B2B.Transactions.Messages;
+using B2B.Transactions.IncomingMessages;
 using B2B.Transactions.Xml.Incoming;
 
 namespace B2B.CimMessageAdapter.Messages
