+++ source/B2B.Transactions.CimMessageAdapter/Messages/MessageParserResult.cs
@@ -14,7 +14,7 @@
 
 using System.Collections.Generic;
 using B2B.CimMessageAdapter.Errors;
-using B2B.Transactions.Messages;
+using B2B.Transactions.IncomingMessages;
 
 namespace B2B.CimMessageAdapter.Messages
 {
