+++ source/B2B.Transactions.IntegrationTests/IncomingMessages/IncomingMessageBuilder.cs
@@ -14,7 +14,6 @@
 
 using System;
 using B2B.Transactions.IncomingMessages;
-using B2B.Transactions.Messages;
 using B2B.Transactions.Transactions;
 
 namespace B2B.Transactions.IntegrationTests.Transactions
